requirements.txt
discord.py==2.0.1
wavelink==0.10.1
python-dotenv==1.0.0
